refactor(main): move runtime diagnostics and screenshot helpers
